+++ src/main/java/cn/tedu/jsdvn2203/csmall/server/service/impl/CategoryServiceImpl.java
@@ -1,5 +1,6 @@
 package cn.tedu.jsdvn2203.csmall.server.service.impl;
 
+import cn.tedu.jsdvn2203.csmall.server.exception.ServiceException;
 import cn.tedu.jsdvn2203.csmall.server.mapper.CategoryMapper;
 import cn.tedu.jsdvn2203.csmall.server.pojo.dto.CategoryAddNewDTO;
 import cn.tedu.jsdvn2203.csmall.server.pojo.dto.CategoryDeleteDTO;
@@ -17,23 +18,37 @@ public class CategoryServiceImpl implements ICategoryService {
     @Autowired
     CategoryMapper categoryMapper;
 
-    public CategoryServiceImpl(){
+    public CategoryServiceImpl() {
         log.info("创建业务逻辑对象.CategoryService");
     }
 
     @Override
     public void addNew(CategoryAddNewDTO categoryAddNewDTO) {
+        String name = categoryAddNewDTO.getName();
+        int count = categoryMapper.countByName(name);
+        if (count > 0) {
+            String message = "添加失败,类别名称[" + name + "]已存在";
+            log.error(message);
+            throw new ServiceException(message);
+        }
+
         Category category = new Category();
-        BeanUtils.copyProperties(categoryAddNewDTO,category);
+        BeanUtils.copyProperties(categoryAddNewDTO, category);
         int rows = categoryMapper.insert(category);
-        log.info("插入成功，受影响的行数：{}",rows);
+        log.info("插入成功，受影响的行数：{}", rows);
     }
 
     @Override
     public void delete(CategoryDeleteDTO categoryDeleteDTO) {
+        Long id = categoryDeleteDTO.getId();
+        int count = categoryMapper.countById(id);
+        if (count == 0) {
+            String message = "删除失败，类别id[" + id + "]不存在";
+            throw new ServiceException(message);
+        }
         Category category = new Category();
-        BeanUtils.copyProperties(categoryDeleteDTO,category);
+        BeanUtils.copyProperties(categoryDeleteDTO, category);
         int rows = categoryMapper.deleteById(category.getId());
-        log.info("删除成功,受影响的行数:{}",rows);
+        log.info("删除成功,受影响的行数:{}", rows);
     }
 }
